Add artifact deletion, typed entities, same_as for memories, fix contradiction job, add query playground
- Artifact deletion: DELETE /v1/knowledge/{id} + UI action; pre-nulls NO ACTION FKs,
  resets memory promotion_status, flags covering digests stale before CASCADE delete;
  requires scope admin; no MCP exposure (deliberate human action only)

- Typed entities: remember now accepts [{name, type}] objects instead of []string;
  entity_type stored verbatim (concept|technology|file|person|service|pr|decision);
  bare strings still accepted as 'concept' for backwards compatibility;
  REST POST /v1/memories updated to match; skill file updated with type table

- Entity auto-extraction: added file:path and @mention patterns to ExtractEntitiesFromMemory

- same_as relations for memory entities: memory/store.go now mirrors knowledge/store.go;
  calls ListEntitiesByCanonical after each upsert and creates same_as to type-siblings

- Fix contradiction job panic: embedding::text scanned into *string but never parsed
  into *pgvector.Vector, leaving pointer nil; fixed by parsing via v.Scan() after each
  row; added nil guard in filterByTopicSimilarity

- Query playground: /ui/query with scope dropdown, search mode, layer checkboxes,
  configurable limit; results table with score, layer, type, content, metadata

+++ internal/db/memories.sql.go
@@ -392,7 +392,7 @@ type RecallMemoriesByCodeVectorRow struct {
 	SourceRef            *string
 	CreatedAt            time.Time
 	UpdatedAt            time.Time
-	VecScore             float32
+	VecScore             int32
 }
 
 func (q *Queries) RecallMemoriesByCodeVector(ctx context.Context, arg RecallMemoriesByCodeVectorParams) ([]*RecallMemoriesByCodeVectorRow, error) {
@@ -673,7 +673,7 @@ type RecallMemoriesByVectorRow struct {
 	SourceRef            *string
 	CreatedAt            time.Time
 	UpdatedAt            time.Time
-	VecScore             float32
+	VecScore             int32
 }
 
 func (q *Queries) RecallMemoriesByVector(ctx context.Context, arg RecallMemoriesByVectorParams) ([]*RecallMemoriesByVectorRow, error) {
